Feat: Added flask blueprints. Routes are in their appropriate directory. Easier for scaling and managing
