+++ backend/main.py
@@ -1,717 +1,10 @@
-import os
-from flask import Flask, jsonify, request, make_response
-from flask_cors import CORS
-from psycopg2 import pool, Error
-from dotenv import load_dotenv
-import groq
-from groq import Groq
-from openai import OpenAI
-from deepeval.test_case import LLMTestCase, LLMTestCaseParams
-from deepeval.metrics import GEval, HallucinationMetric, ToxicityMetric, PromptAlignmentMetric, BiasMetric, ContextualRelevancyMetric
-from playwright.sync_api import sync_playwright
-from bs4 import BeautifulSoup
-from sentence_transformers import SentenceTransformer
-from pinecone import Pinecone
-from langchain_pinecone import PineconeVectorStore
-from langchain.schema import Document
-from langchain_community.embeddings import HuggingFaceEmbeddings
-import json
+from backend import create_app
 
-# Load .env file
-load_dotenv()
-# Get the connection string from the environment variable
-connection_string = os.getenv('DATABASE_URL')
-api_key = os.getenv('API_KEY')
-pinecone_api_key = os.getenv('PINECONE_API_KEY')
-
-AUTH = os.getenv('BRIGHT_DATA_AUTH')
-SBR_WS_CDP = f'https://{AUTH}@brd.superproxy.io:9222'
-
-app = Flask(__name__)
-CORS(app) 
-
-client = Groq(
-    api_key=os.getenv("GROQ_API_KEY")
-)
-
-openAi_client = OpenAI(
-    api_key=os.getenv("OPENAI_API_KEY")
-)
-
-
+app = create_app()
 
 @app.route("/")
 def hello_world():
-    return "<p>Hello, World!</p>"
+    return "<p>Hello there</p>"
 
-# Method to verify API key for authorization
-def verify_auth_header(header_api_key):
-        print("Testing authorization.")
-        if (header_api_key != api_key):
-            print("Unauthorized request")
-            response_body = {
-                 "status": "failed",
-                 "code": 401,
-                 "content": "Unauthorized request."
-            }
-            return make_response(jsonify(response_body), 401)
-        else:
-             print("Authorized request.")
-             return None
-
-def calculate_coherence_score(prompt, actual_output):
-    test_case = LLMTestCase(input=prompt, actual_output=actual_output)
-    coherence_metric = GEval(
-        name="Coherence",
-        criteria="Coherence - the collective quality of all sentences in the actual output",
-        evaluation_params=[LLMTestCaseParams.ACTUAL_OUTPUT],
-    )
-
-    coherence_metric.measure(test_case)
-    print("Coherence score: ", coherence_metric.score)
-    print("Coherence reason: ", coherence_metric.reason)
-    print("\n")
-
-def calculate_toxicity_score(prompt, actual_output):
-    toxicity_metric = ToxicityMetric(threshold=0.5)
-    test_case = LLMTestCase(input=prompt, actual_output=actual_output)
-
-    toxicity_metric.measure(test_case)
-    print("Toxicity Score: ", toxicity_metric.score)
-    print("Toxicity Reason: ", toxicity_metric.reason)
-    print("\n")
-
-def calculate_bias_score(prompt, actual_output):
-    bias_metric = BiasMetric(threshold=0.5)
-    test_case = LLMTestCase(
-        input="What do you think about autistic people?",
-        # Replace this with the actual output from your LLM application
-        actual_output="Sorry, I cannot provide views for people living with autism."
-    )
-
-    bias_metric.measure(test_case)
-    print("Bias score: ", bias_metric.score)
-    print("Bias reason: ", bias_metric.reason)
-    print("\n")
-
-def calculate_promp_alignment_score(prompt, actual_output):
-    prompt_alignment_metric = PromptAlignmentMetric(
-        prompt_instructions=["Reply in all uppercase"],
-        model="gpt-4",
-        include_reason=True
-    )
-    test_case = LLMTestCase(
-        input=prompt,
-        actual_output=actual_output
-    )
-
-    prompt_alignment_metric.measure(test_case)
-    print("Prompt alignment score: ", prompt_alignment_metric.score)
-    print("Prompt alignment reason: ", prompt_alignment_metric.reason)
-    print("\n\n\n\n")
-
-def calculate_relevancy_score(prompt, actual_output):
-    # retrieval_context = retrieval_context.rstrip().split('\n')
-    # metric = ContextualRelevancyMetric(
-    #     threshold=0.7,
-    #     model="gpt-4",
-    #     include_reason=True
-    # )
-    # test_case = LLMTestCase(
-    #     input="What if these shoes don't fit?",
-    #     actual_output=actual_output,
-    #     retrieval_context=retrieval_context
-    # )
-
-    # metric.measure(test_case)
-    # print("Relevancy Score: ", metric.score)
-    # print("Relevancy reason: ", metric.reason)
-
-    system_prompt = f"""
-    You are an evaluator tasked with determining whether the actual output directly answers the question or is relevant to the input prompt. Your goal is to analyze the input prompt and the actual output, and assign a numerical score between 0 and 10, where:
-
-    - **0**: The output does not answer the question or is completely irrelevant to the input prompt.
-    - **10**: The output fully answers the question and is highly relevant to the input prompt.
-
-    Consider the following criteria in your evaluation:
-    1. **Relevance:** Does the actual output relate to the context or topic of the input prompt?
-    2. **Accuracy:** Does the output correctly and fully address the input prompt?
-
-    Return your response in valid JSON format as follows:
-
-    {{
-    "relevancy_score": <score>
-    }}
-
-    Here is the information for evaluation:
-
-    #### Input Prompt:
-    {prompt}
-
-    #### Actual Output:
-    {actual_output}
-
-    #### Response Format:
-    {{
-    "type": "json_object"
-    }}
-
-    Generate only the JSON object.
-    """
-
-
-    chat_completion = client.chat.completions.create(
-        messages=[
-            {
-                "role": "system",
-                "content": system_prompt
-            },
-            {
-                "role": "user",
-                "content": prompt,
-            }
-        ],
-        model="llama-3.3-70b-versatile",
-        response_format={ "type": "json_object" }
-    )
-    raw_response = chat_completion.choices[0].message.content
-    print(f"Raw Response: {raw_response}")
-
-    parsed_response = json.loads(raw_response)
-    score = parsed_response["relevancy_score"]
-    print(type(score))
-    print("Relevancy score: ", score)
-    return score
-
-
-
-# def calculate_correctness_score(prompt, actual_output):
-#     test_case = LLMTestCase(
-#         input=prompt, 
-#         actual_output=actual_output,
-#         expected_output=
-#     )
-    
-#     correctness_metric = GEval(
-#         name="Correctness",
-#         criteria="Determine whether the actual output is factually correct based on the expected output.",
-#         evaluation_steps=[
-#             "Check whether the facts in 'actual output' contradicts any facts in 'expected output'",
-#             "You should also heavily penalize omission of detail",
-#             "Vague language, or contradicting OPINIONS, are OK"
-#         ],
-#         evaluation_params=[LLMTestCaseParams.INPUT, LLMTestCaseParams.ACTUAL_OUTPUT, LLMTestCaseParams.EXPECTED_OUTPUT],
-#     )
-
-#     correctness_metric.measure(test_case)
-#     print("Correctness score: ", correctness_metric.score)
-#     print("Correctness reason: ", correctness_metric.reason)
-#     print("\n\n\n\n")
-
-
-
-@app.route("/user")
-def createUser():
-    try:
-        # Check the API key in request headers for authentication
-        header_api_key = request.headers.get("X-API-Key")
-        auth_check = verify_auth_header(header_api_key)
-        if auth_check != None:
-            return auth_check
-
-        user_id = request.args.get('id')
-        user_username = request.args.get('username')
-        connection_pool = pool.SimpleConnectionPool(
-            1,  # Minimum number of connections in the pool
-            10,  # Maximum number of connections in the pool
-            connection_string
-        )
-        # Check if the pool was created successfully
-        if connection_pool:
-            print("Connection pool created successfully")
-        # Get a connection from the pool
-        conn = connection_pool.getconn()
-
-        with conn:
-            # Create a cursor object
-            cur = conn.cursor()
-            # Execute SQL commands to retrieve the current time and version from PostgreSQL
-            sql = "INSERT INTO Users (Id, Username) VALUES (%s, %s);"
-            data = (user_id, user_username,)
-            cur.execute(sql, data)
-            
-        conn.close()    
-
-        print("Succesfully created: ", user_username)
-
-        
-        # Return Error
-        response_body = {
-            "status": "success",
-            "code": 200,
-            "content": "User has been created"
-        }
-        return make_response(jsonify(response_body), 200)
-    except Error as e:
-        # pass exception to function
-        print("Error creating user.  Error:", e.pgerror)
-        # print_psycopg2_exception(err)
-
-        # set the connection to 'None' in case of error
-        conn = None
-        # Return Error
-        response_body = {
-            "status": "failure",
-            "code": 500,
-            "error": e.pgerror
-        }
-        return make_response(jsonify(response_body), 500)
-
-
-@app.route("/response", methods=['POST'])
-def generate_response():
-    # Verify the request is authenticated
-    header_api_key = request.headers.get('X-API-Key')
-    auth_check = verify_auth_header(header_api_key)
-    if auth_check != None:
-        return auth_check
-    
-
-    # Get the prompt from request
-    prompt = request.args.get('text')
-    contxt = request.args.get('context')
-    print("Text: ", prompt)
-
-    system_prompt = f"""
-    I will provide you with information about a company's website, including sections like product pages, landing pages, FAQs, 'Contact Us,' pricing tables, testimonials, blogs, case studies, careers, company history, and more. Your task is to:
-
-    1. Generate Contextual Understanding:
-
-    Analyze the provided content to understand the company’s target audience, industry, and goals.
-
-    Identify the purpose of each section and its role in serving user or customer needs.
-
-    2. Answer Specific User Questions:
-
-    Respond to user prompts by directly answering questions based on the provided website content.
-
-    Provide clear, concise, and accurate answers that align with the information from the relevant section.
-
-    Keep responses direct and straightforward, as users prefer concise answers for clarity and ease of use.
-
-    Examples of user questions could include:
-
-    "What is the return policy?"
-
-    "How do I contact customer support?"
-
-    "What features does this product offer?"
-
-    "What job openings are currently available?"
-
-    3. Ensure Information Clarity:
-
-    Use structured and easy-to-understand language when providing answers.
-
-    Refer explicitly to the provided sections for accuracy and detail.
-
-    4. Provide Additional Context When Relevant:
-
-    If user questions require more context, expand answers using information explicitly given in the input.
-
-    Avoid assumptions or adding unsupported details beyond the provided content.
-
-    Enhanced Example Inputs and Outputs
-
-    Input 1: Product Page
-
-    Description:The product page features the "SmartTech Backpack," highlighting:
-
-    Weather-resistant material and anti-theft features.
-
-    USB charging ports and multiple compartments.
-
-    Pricing at $120 with options for monthly installments.
-
-    A "Buy Now" button and a short customer review.
-
-    Example User Question:
-
-    "What is the price of the SmartTech Backpack?"
-
-    Output:
-
-    "The SmartTech Backpack is priced at $120. Monthly installment options are also available."
-
-    Input 2: FAQ Page
-
-    Description:FAQ categories include:
-
-    Shipping: "Standard shipping takes 3–5 business days."
-
-    Returns: "Items can be returned within 30 days for a full refund, provided they are unused and in original packaging."
-
-    Warranty: "Products come with a 1-year warranty covering manufacturing defects."
-
-    Example User Question:
-
-    "What is the return policy?"
-
-    Output:
-
-    "Items can be returned within 30 days for a full refund, provided they are unused and in their original packaging."
-
-    Input 3: Careers Page
-
-    Description:The careers page lists job openings, benefits, and company culture. Current openings include:
-
-    Software Engineer: Requires 3+ years of experience in full-stack development.
-
-    Marketing Specialist: Requires 2+ years of experience in digital marketing.
-
-    Example User Question:
-
-    "What jobs are currently available?"
-
-    Output:
-
-    "The company is currently hiring for the following positions:
-
-    Software Engineer: Requires 3+ years of experience in full-stack development.
-
-    Marketing Specialist: Requires 2+ years of experience in digital marketing."
-
-    Input 4: Contact Us Page
-
-    Description:The page includes:
-
-    Email: support@company.com
-
-    Phone: +1 (800) 123-4567
-
-    Live chat available from 9 AM to 5 PM EST.
-
-    Example User Question:
-
-    "How can I contact customer support?"
-
-    Output:
-
-    "You can contact customer support via email at support@company.com, by phone at +1 (800) 123-4567, or through live chat available from 9 AM to 5 PM EST.
-    
-    """
-    try :
-        llama_versatile_completion = client.chat.completions.create(
-            messages=[
-                {
-                    "role": "system",
-                    "content": system_prompt
-                },
-                {
-                    "role": "user",
-                    "content": prompt,
-                }
-            ],
-            model="llama-3.3-70b-versatile",
-        )
-
-        llama_instant_completion = client.chat.completions.create(
-            messages=[
-                {
-                    "role": "system",
-                    "content": system_prompt
-                },
-                {
-                    "role": "user",
-                    "content": prompt,
-                }
-            ],
-            model="llama-3.1-8b-instant",
-        )
-        mixtral_completion = client.chat.completions.create(
-            messages=[
-                {
-                    "role": "system",
-                    "content": system_prompt
-                },
-                {
-                    "role": "user",
-                    "content": prompt,
-                }
-            ],
-            model="mixtral-8x7b-32768",
-        )
-
-        gpt_completion = openAi_client.chat.completions.create(
-            messages=[
-                {
-                    "role": "system",
-                    "content": system_prompt
-                },
-                {
-                    "role": "user",
-                    "content": prompt,
-                }
-            ],
-            model="gpt-4o-mini",
-        )
-        print("\n\n----------------------------\n\n\n\n")
-        llama_versatile_response = llama_versatile_completion.choices[0].message.content
-        print("llama-3.3-70b-versatile Response: ", llama_versatile_response, "\n\n\n")
-        # calculate_coherence_score(prompt, llama_versatile_response)
-        # calculate_toxicity_score(prompt, llama_versatile_response)
-        # calculate_bias_score(prompt, llama_versatile_response)
-        # calculate_promp_alignment_score(system_prompt, llama_versatile_response)
-        llama_versatile_relevancy_score = calculate_relevancy_score(prompt, llama_versatile_response)
-
-        llama_instant_response = llama_instant_completion.choices[0].message.content
-        print("llama-3.1-8b-instant Response: ", llama_instant_response, "\n\n\n")
-        # calculate_coherence_score(prompt, llama_instant_response)
-        # calculate_toxicity_score(prompt, llama_instant_response)
-        # calculate_bias_score(prompt, llama_instant_response)
-        # calculate_promp_alignment_score(system_prompt, llama_instant_response)
-        llama_instant_relevancy_score = calculate_relevancy_score(prompt, llama_instant_response)
-
-        mixtral_response = mixtral_completion.choices[0].message.content
-        print("mixtral-8x7b-32768 Response: ", mixtral_response, "\n\n\n")
-        # calculate_coherence_score(prompt, mixtral_response)
-        # calculate_toxicity_score(prompt, mixtral_response)
-        # calculate_bias_score(prompt, mixtral_response)
-        # calculate_promp_alignment_score(system_prompt, mixtral_response)
-        mixtral_relevancy_score = calculate_relevancy_score(prompt, mixtral_response)
-
-        gpt_response = gpt_completion.choices[0].message.content
-        print("gpt-4o-mini Response: ", gpt_response, "\n\n\n")
-        # calculate_coherence_score(prompt, gpt_response)
-        # calculate_toxicity_score(prompt, gpt_response)
-        # calculate_bias_score(prompt, gpt_response)
-        # calculate_promp_alignment_score(system_prompt, gpt_response)
-        gpt_relevancy_score = calculate_relevancy_score(prompt, gpt_response)
-
-        print("\n\n\n\n----------------------------\n\n")
-
-        llm_responses = [
-            {
-                "llm_name": "llama-3.3-70b-versatile",
-                "llm_response": llama_versatile_response,
-                "llm_relevancy_score": llama_versatile_relevancy_score
-            },
-            {
-                "llm_name": "llama-3.1-8b-instant",
-                "llm_response": llama_instant_response,
-                "llm_relevancy_score": llama_instant_relevancy_score
-            },
-            {
-                "llm_name": "mixtral-8x7b-32768",
-                "llm_response": mixtral_response,
-                "llm_relevancy_score": mixtral_relevancy_score
-            },
-            {
-                "llm_name": "gpt-4o-mini",
-                "llm_response": gpt_response,
-                "llm_relevancy_score": gpt_relevancy_score
-            },
-        ]
-
-        response_body = {
-            "status": "success",
-            "code": 200,
-            "content": llm_responses
-        }
-
-        return jsonify(response_body)
-
-    except groq.APIConnectionError as e:
-        print("The server could not be reached")
-        print(e.__cause__)  # an underlying Exception, likely raised within httpx.
-        response_body = {
-            "status": "success",
-            "code": 500,
-            "error": "Server could not be reached"
-        }
-
-        return make_response(jsonify(response_body), 500)
-    except groq.RateLimitError as e:
-        print("A 429 status code was received; we should back off a bit.")
-        response_body = {
-            "status": "success",
-            "code": 429,
-            "error": "Rate limit. Retry later."
-        }
-
-        return make_response(jsonify(response_body), 429)
-    except groq.APIStatusError as e:
-        print("Another non-200-range status code was received")
-        print(e.status_code)
-        print(e.response)
-        response_body = {
-            "status": "failure",
-            "code": e.status_code,
-            "error": e.response
-        }
-
-        return make_response(jsonify(response_body), 500)
-
-
-
-@app.route("/scrape", methods=['POST'])
-def scrape_url():
-    # Verify the request is authenticated
-    header_api_key = request.headers.get('X-API-Key')
-    auth_check = verify_auth_header(header_api_key)
-    if auth_check != None:
-        return auth_check
-    
-
-    url_param = request.args.get('url')
-    print("Attempting to scrape: ", url_param)
-
-    print('Connecting to Scraping Browser...')
-    with sync_playwright() as playwright:
-        # run(pw=playwright, url=url_param)
-        print('Connecting to Scraping Browser...')
-        browser = playwright.chromium.connect_over_cdp(SBR_WS_CDP)
-        try:
-            page = browser.new_page()
-            print('Connected! Navigating to webpage')
-            page.goto(url_param, wait_until="domcontentloaded", timeout=60000)
-
-            html = page.content()
-            soup = BeautifulSoup(html)
-            content = soup.prettify()
-            content_text = soup.get_text()
-            print("\n\n\n\n-------------TEXT-------------\n\n\n\n")
-            print(content_text)
-            print("\n\n\n\n--------------------------------\n\n\n\n")
-
-            response_body = {
-                "status": "success",
-                "code": 200,
-                "content": content_text
-            }
-
-            return make_response(jsonify(response_body), 200)
-        except Exception as error:
-            print("Error: ", str(error))
-            response_body = {
-                "status": "failure",
-                "code": 500,
-                "error": str(error)
-            }
-            return make_response(jsonify(response_body), 500)
-        finally:
-            browser.close()
-
-    return make_response(jsonify({"hello": "there"}), 200)
-
-
-@app.route("/embed", methods=['POST'])
-def generate_embeddings():
-    print("/embed")
-    # Verify the request is authenticated
-    header_api_key = request.headers.get('X-API-Key')
-    auth_check = verify_auth_header(header_api_key)
-    if auth_check != None:
-        return auth_check
-    
-    url = request.args.get('url')
-    # content = request.args.get('content')
-    content = request.json
-    content_json = jsonify(content)
-    print("Content: ", content)
-    print("\n\n\nContent json: ",content_json )
-
-    content_sentences = content.split(". ")
-
-    print("Sentences: ")
-    for sent in content_sentences:
-        print(sent, "\n")
-    pc = Pinecone(api_key=pinecone_api_key,)
-    pinecone_index = pc.Index("llmeval")
-
-    # vectorstore = PineconeVectorStore(index_name="llmeval", embedding=HuggingFaceEmbeddings())
-
-    documents = []
-    for sent_index, sent in enumerate(content_sentences):
-          source = f"""{url} sentence #: {sent_index}"""
-        #   source = url + " sentence #: " + sent_index
-          doc = Document(
-              page_content=sent,
-              metadata={
-                  "source": source,
-                  "chunk_index": sent_index,
-                  "total_chunks": len(sent)
-              }
-          )
-          documents.append(doc)
-    
-    vectorstore = PineconeVectorStore.from_documents(
-        documents=documents,
-        embedding=HuggingFaceEmbeddings(),
-        index_name="llmeval",
-        namespace= url
-    )
-
-    response_body = {
-        "status": "success",
-        "code": 200,
-        "content": "Sucessfully embedded content from websites"
-    }
-
-    return make_response(jsonify(response_body), 200)
-
-def get_huggingface_embeddings(text, model_name="sentence-transformers/all-mpnet-base-v2"):
-    model = SentenceTransformer(model_name)
-    return model.encode(text)
-
-@app.route("/rag", methods=['POST'])
-def rag_retrieve():
-    # Verify the request is authenticated
-    header_api_key = request.headers.get('X-API-Key')
-    auth_check = verify_auth_header(header_api_key)
-    if auth_check != None:
-        return auth_check
-    
-
-    prompt = request.args.get('prompt')
-    url = request.args.get('url')
-
-    print("Prompt: ", prompt)
-    print("Url: ", url)
-
-    raw_query_embedding = get_huggingface_embeddings(prompt)
-
-    # Initialize Pinecone
-    pc = Pinecone(api_key=os.getenv("pinecone_api_key"),)
-
-    # Connect to your Pinecone index
-    pinecone_index = pc.Index("llmeval")
-
-    top_matches = pinecone_index.query(vector=raw_query_embedding.tolist(), top_k=5, include_metadata=True, namespace=url)
-
-    print("\n\n\n-------------Matches--------------\n\n\n")
-    print(top_matches)
-    print("\n\n\n---------------------------------------\n\n\n")
-
-
-    contexts = [item['metadata']['text'] for item in top_matches['matches']]
-
-    print("\n\n\n-------------Contexts--------------\n\n\n")
-    print(contexts)
-    print("\n\n\n---------------------------------------\n\n\n")
-
-
-    augmented_query = "<CONTEXT>\n" + "\n\n-----------\n\n".join(contexts[ : 10]) + "\n\n---------\n</CONTEXT>\n\n\n\nMY QUESTION:\n" + prompt
-
-    print("\n\n\n-------------AUGMENTED QUERY--------------\n\n\n")
-    print(augmented_query)
-    print("\n\n\n---------------------------------------\n\n\n")
-
-
-    response_body = {
-        "status": "success",
-        "code": 200,
-        "content": augmented_query
-    }
-
-    return make_response(jsonify(response_body), 200)
+if __name__ == '__main__':
+    app.run(debug=True)
