+++ backend/rag/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('rag', __name__)
+
+from backend.rag import routes
