+++ backend/scrape/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('scrape', __name__)
+
+from backend.scrape import routes
