+++ backend/scrape/routes.py
@@ -1,0 +1,56 @@
+from flask import jsonify, request, make_response
+from backend.scrape import bp
+from backend.utils import verify_auth_header
+from playwright.sync_api import sync_playwright
+from bs4 import BeautifulSoup
+from backend.config import Config
+
+AUTH = Config.BRIGHT_DATA_AUTH
+SBR_WS_CDP = Config.SBR_SW_CDP
+
+@bp.route("/", methods=['POST'])
+def scrape_url():
+    # Verify the request is authenticated
+    header_api_key = request.headers.get('X-API-Key')
+    auth_check = verify_auth_header(header_api_key)
+    if auth_check != None:
+        return auth_check
+    
+    url_param = request.args.get('url')
+    print("Attempting to scrape: ", url_param)
+
+    print('Connecting to Scraping Browser...')
+    with sync_playwright() as playwright:
+        # run(pw=playwright, url=url_param)
+        print('Connecting to Scraping Browser...')
+        browser = playwright.chromium.connect_over_cdp(SBR_WS_CDP)
+        try:
+            page = browser.new_page()
+            print('Connected! Navigating to webpage')
+            page.goto(url_param, wait_until="domcontentloaded", timeout=60000)
+
+            html = page.content()
+            soup = BeautifulSoup(html)
+            content = soup.prettify()
+            content_text = soup.get_text()
+            print("\n\n\n\n-------------TEXT-------------\n\n\n\n")
+            print(content_text)
+            print("\n\n\n\n--------------------------------\n\n\n\n")
+
+            response_body = {
+                "status": "success",
+                "code": 200,
+                "content": content_text
+            }
+
+            return make_response(jsonify(response_body), 200)
+        except Exception as error:
+            print("Error: ", str(error))
+            response_body = {
+                "status": "failure",
+                "code": 500,
+                "error": str(error)
+            }
+            return make_response(jsonify(response_body), 500)
+        finally:
+            browser.close()
